Support for installation on cents.

+++ deployment_scripts/experiment_manager/install.sh
@@ -5,21 +5,36 @@ source `dirname $0`/../utils.sh
 export RAILS_ENV=production
 
 info "Packages installation"
-execute sudo apt-get update
-execute sudo apt-get install -y curl zip git sysstat r-base-core redis-server
+if [ -a /etc/redhat-release ]; then
+	execute sudo yum install -y git curl R zip sysstat redis golang wget
+
+	info "GSISSH installation - getting Globus repo"
+	execute wget http://toolkit.globus.org/ftppub/gt6/installers/repo/globus-toolkit-repo-latest.noarch.rpm -O /tmp/globus-toolkit-repo-latest.noarch.rpm
+
+	info "GSISSH installation - installing Globus repo"
+	execute sudo rpm -ivh /tmp/globus-toolkit-repo-latest.noarch.rpm && rm /tmp/globus-toolkit-repo-latest.noarch.rpm
+
+	info "GSISSH installation - installing gsi-openssh-clients"
+	execute sudo yum install -y gsi-openssh-clients
+else
+	if [ -a /etc/os-release ] && [ `cat /etc/os-release | grep ID=ubuntu | wc -l` == '1' ]; then
+		execute sudo apt-get update
+		execute sudo apt-get install -y curl zip git sysstat r-base-core redis-server
+
+		info "GSISSH installation - getting Globus repo"
+		execute wget http://toolkit.globus.org/ftppub/gt6/installers/repo/globus-toolkit-repo_latest_all.deb -O /tmp/globus-toolkit-repo_latest_all.deb
+
+		info "GSISSH installation - installing Globus repo"
+		execute sudo dpkg -i /tmp/globus-toolkit-repo_latest_all.deb && rm /tmp/globus-toolkit-repo_latest_all.deb
+
+		info "GSISSH installation - installing gsi-openssh-clients"
+		execute sudo apt-get update
+		execute sudo apt-get install -y gsi-openssh-clients libglobus-gssapi-gsi4
+	fi	
+fi
 
 info "Go to SCALARM_ROOT: $SCALARM_ROOT"
 execute cd $SCALARM_ROOT
-
-info "GSISSH installation - getting Globus repo"
-execute wget http://toolkit.globus.org/ftppub/gt6/installers/repo/globus-toolkit-repo_latest_all.deb -O /tmp/globus-toolkit-repo_latest_all.deb
-
-info "GSISSH installation - installing Globus repo"
-execute sudo dpkg -i /tmp/globus-toolkit-repo_latest_all.deb && rm /tmp/globus-toolkit-repo_latest_all.deb
-
-info "GSISSH installation - installing gsi-openssh-clients"
-execute sudo apt-get update
-execute sudo apt-get install -y gsi-openssh-clients libglobus-gssapi-gsi4
 
 info "Installing PLGrid certificates (for gsissh)"
 execute $SCRIPT_PATH/../install_plgrid_certs.sh
